01_etl/requirements.txt
elasticsearch==7.7.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
redis==4.3.1
pydantic==1.9.0